% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ChangeDetection.r
\name{CD_list}
\alias{CD_list}
\title{Change Detection}
\usage{
CD_list(img.list, ...)
}
\arguments{
\item{img.list}{file path to images.}

\item{...}{Arguments passed to CD_single().}
}
\value{
Classification result. Here work is in progess...
}
\description{
Detect changes using image differencing for a list of images.
Includes parallel processing.
}
