% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ChangeDetection.r
\name{CD_single}
\alias{CD_single}
\title{Detect change.difs using image differencing}
\usage{
CD_single(now, old, min = 30/255, max = 220/255, extend = NULL)
}
\arguments{
\item{now}{Path to first image}

\item{old}{Path to second image}

\item{min}{Threshold for positive classification}

\item{max}{Threshold for positive classification}

\item{extend}{DEPECATED!
Used to crop images. Has been moved to a seperate preprocess step.}
}
\value{
Classification result. Here work is in progess...
}
\description{
in a single image.
}
