% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ChangeDetection.r
\name{ChangeDetection}
\alias{ChangeDetection}
\title{Detect changes using image differencing}
\usage{
ChangeDetection(now, old, min = 30/255, max = 220/255, extend = NULL)
}
\arguments{
\item{now}{Path to first image}

\item{old}{Path to second image}

\item{min}{Threshold for positive classification}

\item{max}{Threshold for positive classification}

\item{extend}{DEPECATED!
Used to crop images. Has been moved to a seperate preprocess step.}
}
\value{
Classification result. Here work is in progess...
}
\author{
Jeroen Staab
}
