% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{fun_Aggregation}
\alias{fun_Aggregation}
\title{Aggregate Time-series}
\usage{
fun_Aggregation(Timestamp, Variable, T_scale = "hour")
}
\arguments{
\item{T_scale}{Timeinterval. See ?lubridate::floor_date()}
}
\value{
Dataframe including summed and mean variable per Timeinterval
}
\description{
Aggregate Time-series
}
