% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{GTD_list}
\alias{GTD_list}
\title{Sample Ground Truth Data}
\usage{
GTD_list(img.list)
}
\arguments{
\item{img.list}{file path to image, also known as `now`.}
}
\value{
numeric vector with number of persons.
}
\description{
Manually asses number of persons in multiple images.
}
