% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_gtd.R
\name{GTD_single}
\alias{GTD_single}
\title{Sample Ground Truth Data}
\usage{
GTD_single(img)
}
\arguments{
\item{img}{file path to image, also known as `now`.}
}
\value{
numeric vector with number of persons.
}
\description{
Manually asses number of persons in a single image.
}
\author{
Jeroen Staab
}
