% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_gtd.R
\name{GTD_truePositives}
\alias{GTD_truePositives}
\title{Benchmark Pedestrian Detection}
\usage{
GTD_truePositives(GTD, PRD)
}
\arguments{
\item{GTD}{Numeric vector of Ground-Truth-Data, as returned by GTD_list()}

\item{PRD}{Numeric vector of prediction values, as returned by hog_list()
and yolo_list()}
}
\value{
FalseNegative, TruePositives, FalsePositive, MissRate,
FalsePositvesPerWindow
}
\description{
Accuracy Assesment for Object-Based Classifiers as in DALAL
etal. 2005 p. 888
}
