% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_hog.R
\name{hog}
\alias{hog}
\title{Detect pedestrians using HOGDescriptor}
\usage{
hog(img.folder)
}
\arguments{
\item{img.folder}{Path to (preprocessed) image archive}

\item{winStride}{Not implemented yet!}

\item{padding}{Not implemented yet!}

\item{scale}{Not implemented yet!}
}
\value{
Numeric vector with number of detected persons.
}
\description{
Detect objects using HOG+SVM (implemented in OpenCV) in all Files/Images of 'path'
}
\details{
Python and OpenCV have to be installed. Tested on Linux only.

Further ideas:
[A] Add more 'hog.detectMultiScale' parameters: winStride=(4, 4), padding=(8, 8), scale=1.05)
[B] Save predictions.png to a folder
}
\author{
Jeroen Staab
}
