% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_hog.R
\name{hog_install}
\alias{hog_install}
\title{How to install HOG-Descriptor?}
\usage{
?hog_install()
}
\description{
hog_list() depends on a functional OpenCV installation.
This is how I installed it on the LSFE workstation (Linux). OS-specific

OpenCV: \code{sudo apt install python-opencv}

Package Manager: \code{sudo apt install python-pip}

HOG Dependency: \code{pip install imutils}

CUDA GPU: \code{sudo apt-get install nvidia-cuda-dev nvidia-cuda-toolkit nvidia-nsight}
}
\seealso{
http://docs.opencv.org/trunk/df/d65/tutorial_table_of_content_introduction.html
}
