% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{JPEG_grayscale}
\alias{JPEG_grayscale}
\title{Convert RGB to Grayscale}
\usage{
JPEG_grayscale(img, red = 1/3, green = 1/3, blue = 1/3)
}
\arguments{
\item{img}{Three layered raster object.}

\item{red}{Calibration weight for red.}

\item{green}{Calibration weight for green.}

\item{blue}{Calibration weight for blue.}
}
\value{
Singe layer raster object.
}
\description{
Convert RGB img to Grayscale. Default is mixing the three
bands equaly. Use camera specific weights if possible.
Stardot red=0.3 green=0.59 blue=0.11
}
