% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{JPEG_histStrecht}
\alias{JPEG_histStrecht}
\title{Histogram Stretching}
\usage{
JPEG_histStrecht(img)
}
\arguments{
\item{img}{A raster object.}
}
\value{
same as input, but ranged between 0 and 1 (nummeric).
}
\description{
Stretch values between 0 and 1 as in JPEG convention.
Attention, use this function for plotting only (highlights contrast). But
further processing (i.e. Change Detection) may be limited due altered values.
}
