% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{plotJPEG}
\alias{plotJPEG}
\title{plotJPEG()}
\usage{
plotJPEG(img, main = NULL)
}
\arguments{
\item{img}{A raster object.}

\item{main}{Optional. An overall title for the plot.}
}
\description{
Simply plot an image.
}
\details{
Images can be loaded as raster object using `JPEG::readJPEG()`.
}
\author{
Jeroen Staab
}
