% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{ROI_draw}
\alias{ROI_draw}
\title{Inspect a Region of Interest}
\usage{
ROI_draw(img)
}
\arguments{
\item{img}{A raster object.}
}
\value{
numeric dataframe with digital numbers of selected pixels.
}
\description{
Draw a region of interest.
}
\details{
Insepct a region of interest by drawing a polygone.
See OS-specific ?locator() for how to finish drawing. Minimum three vertex
points are required.

roi <- InspectROI(jpeg::readJPEG("../Testbild.jpg"))

To visualize roi use histROI() or get it's stats with summary().

histROI(roi)
}
\seealso{
\code{\link{histROI}}
}
