% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{ROI_hist}
\alias{ROI_hist}
\title{Inspect a Region of Interest}
\usage{
ROI_hist(roi)
}
\arguments{
\item{roi}{A numeric dataframe as returned by InspectROI().}
}
\description{
Ggplot histogramm of a region of interest.
}
