% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_yolo.R
\name{yolo}
\alias{yolo}
\title{Detect people usind YOLOv2}
\usage{
yolo(now, logfile = "yolo_detections.txt")
}
\arguments{
\item{now}{Absolute filepath to image.}

\item{logfile}{Relative filepath to where to store detailed list of
classification results.}
}
\value{
Numeric number of detected persons.
}
\description{
Detect objects using YOLO+CNN (Linux C++), in a single image.
}
\details{
Further ideas:
- Skip saving predictions.png
- Wrap detection into list-based processing
   ./darknet detect cfg/yolo.cfg yolo.weights < listofabsolutepaths.txt
- Use RCCP to wrap YOLO into R.
}
