% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_yolo.R
\name{yolo_install}
\alias{yolo_install}
\title{Install YOLO Automatically}
\usage{
yolo_install(yolo.inst)
}
\arguments{
\item{yolo.inst}{directory for installation. Will be created}
}
\description{

}
\details{
this function wrapped the install procedure (1-3) while renaming
the directory from `darknet` to `basename(yolo.inst)` and additional run a
test to check whether installation succeeded.

1. 'git clone https://github.com/pjreddie/darknet'

2. 'cd darknet'

3. 'make'

during installation `Makefile` will be opened, to finetune the
installation, eg. truning on multithreading or GPU processing.

after successfull installation it will place `yolo.inst` in
`paste0(system.file(package = "wuepix"), "/exec/yolo_inst.txt")`
}
