% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_yolo.R
\name{yolo_list}
\alias{yolo_list}
\title{Detect people using YOLO in multiple images.}
\usage{
yolo_list(img.list, logfile = "yolo_detections.txt")
}
\arguments{
\item{img.list}{(Absolute) filepath to image, also known as `now`.}

\item{logfile}{Relative filepath to where to store detailed list of
classification results.}
}
\value{
Numeric number of detected persons.
}
\description{
Detect objects using YOLO+CNN (Linux C++), in multiple images.
}
\details{
Unfortuanatly it is not possible to store the prediction images.
}
