% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_yolo.R
\name{yolo_Read}
\alias{yolo_Read}
\title{Read YOLO Output File}
\usage{
yolo_Read(file = "yolo_detections.txt")
}
\arguments{
\item{file}{path to output "yolo_detections.txt" file.}
}
\description{
Read and clean YOLO output file, as saved to working directory
by yolo_list()
}
