% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_yolo.R
\name{yolo_single}
\alias{yolo_single}
\title{Detect people using YOLO in a single image.}
\usage{
yolo_single(img, logfile = "yolo_detections.txt",
  predictions = "yolo_predictions/")
}
\arguments{
\item{img}{(Absolute) filepath to image, also known as `now`.}

\item{logfile}{Relative filepath to where to store detailed list of
classification results.}

\item{predictions}{dirpath to where to store prediction images}
}
\value{
Numeric number of detected persons.
}
\description{
Detect objects using YOLO+CNN (Linux C++), in a single image.
}
\details{
Single processing allows storing `predictions` (images with
bounding boxes). Since these can be very insightful, you migth want to
`sapply()` this function instead of `yolo_list()`. However because then
the wights have to be loaded repetivly (~10 seconds) this slows down
processing.

It's recomendended avoid spaces in the paths (also in working
directory).

Further ideas:
- Skip saving predictions.png
- Use RCCP to wrap YOLO into R.
}
