% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_yolo.R
\name{yolo_update}
\alias{yolo_update}
\title{Update YOLO}
\usage{
yolo_update(yolo.inst)
}
\arguments{
\item{yolo.inst}{directory of YOLO installation.}
}
\description{
Update YOLO
}
\details{
since YOLO is under active development this function wraps the
update procedure (1-2) and test it.

1. 'git pull'

2. 'make'

during installation `Makefile` will be opened, to fine tune the
installation, e.g. turning on multithreading or GPU processing.

after successfull update it will place `yolo.inst` in
`paste0(system.file(package = "wuepix"), "/exec/yolo_inst.txt")`
}
