% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_yolo.R
\name{yolo}
\alias{yolo}
\title{Detect people usind YOLOv2}
\usage{
yolo(img, logfile = "yolo_detections.txt",
  predictions = "yolo_predictions/")
}
\arguments{
\item{img}{(Absolute) filepath to image, also known as `now`.}

\item{logfile}{Relative filepath to where to store detailed list of
classification results.}

\item{predictions}{dirpath to where to store predictions}
}
\value{
Numeric number of detected persons.
}
\description{
Detect objects using YOLO+CNN (Linux C++), in a single image.
}
\details{
It's recomendended avoid spaces in the paths (also in working
directory).

Further ideas:
- Skip saving predictions.png
- Wrap detection into list-based processing
   ./darknet detect cfg/yolo.cfg yolo.weights < listofabsolutepaths.txt
- Use RCCP to wrap YOLO into R.
}
