% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_yolo.R
\name{yolo_list}
\alias{yolo_list}
\title{Object Detection using YOLO}
\usage{
yolo_list(img.list, logfile = "yolo_detections.txt")
}
\arguments{
\item{img.list}{file path to image, also known as `now`.}

\item{logfile}{file path to where to store detailed list of
classification results.}
}
\value{
numeric number of detected persons.
}
\description{
detect people using YOLO+CNN (Linux C++), in multiple images.
Unfortunately it is not possible to store the predictions here, but it is
significant faster on large image archives.
}
\seealso{
\code{\link{yolo_single}}
}
