% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_yolo.R
\name{yolo_single}
\alias{yolo_single}
\title{Object Detection using YOLO}
\usage{
yolo_single(img, logfile = "yolo_detections.txt",
  predictions = "yolo_predictions/")
}
\arguments{
\item{img}{file path to image, also known as `now`.}

\item{logfile}{file path to where to store detailed list of classification
results.}

\item{predictions}{dir path to where to store prediction images}
}
\value{
numeric number of detected persons.
}
\description{
detect people using YOLO+CNN (Linux C++), in a single image.
}
\details{
depends on a working YOLO installation! See `yolo_install()` and
rerun `yolo_update()` after updating this package (Places yolo.inst in
package directory)

single processing allows storing `predictions` (images with
bounding boxes). Since these can be very insightful, you might want to
`sapply()` this function instead of `yolo_list()`. However because then the
wights have to be loaded repetitively (~10 seconds) this slows down
processing.

it's recommended avoid spaces in the paths (also in working
directory).

an idea for further work on this package would be to actually wrap
YOLO into R (e.g. using Rccp).
}
\examples{
yolo_single(img)
sapply(img.list, yolo_single)

}
\references{
\insertRef{redmon2016yolo9000}{wuepix}
\url{https://pjreddie.com/darknet/yolo/}
}
\author{
Jeroen Staab
}
